pyproject.toml
[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "pytest-cov>=5.0.0",
]

[dependency-groups]
dev = [
    "anyio>=4.4.0",
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short -m 'not db_migration'"
markers = [
    "unit: pure unit tests with no DB or HTTP dependencies",
    "auth: authentication and authorization tests",
    "api: HTTP/API tests",
    "db: tests that require a disposable PostgreSQL database",
    "db_migration: tests that run Alembic migrations against a live PostgreSQL instance (run with: pytest -m db_migration)",
]

[project]
name = "template-service"
version = "0.1.0"
description = "Multi-tenant Template Service for onboarding and verification workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "greenlet>=3.3.1",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.13.0",
    "simpleeval>=1.0.0",
    "sqlmodel>=0.0.34",
    "temporalio>=1.9.0",
    "uuid7>=0.1.0",
    "python-jose[cryptography]>=3.3.0",
]
